backend/requirements.txt
# ================================
# ANCLORA Nexus - BACKEND DEPENDENCIES
# Optimizado - Fase 2 de consolidación
# ================================

# === CORE WEB FRAMEWORK ===
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
Werkzeug==3.1.3

# === DATABASE ===
SQLAlchemy==2.0.23

# === DOCUMENT PROCESSING ===
python-docx==1.2.0
fpdf2==2.8.4
pypdf==6.0.0
lxml==6.0.0
Pillow==11.3.0

# === SECURITY ===
bcrypt==4.3.0
PyJWT==2.10.1

# === TEXT PROCESSING ===
ftfy==6.2.0
chardet==5.2.0

# === ENVIRONMENT & CONFIG ===
python-dotenv==1.0.0

# === MONITORING & METRICS ===
prometheus-flask-exporter==0.23.2

# === PRODUCTION SERVER ===
gunicorn==21.2.0

# === TESTING (Development) ===
pytest==8.3.3
pytest-flask==1.3.0
pytest-cov==6.0.0


